Optionally use a per-package filesystem for `PackageBuilder`
Currently, we have the ability to provide a custom file system to the whole workspace and we partially use git-backed file system during package resolution. This change aims to allow using a custom file system when running `PackageBuilder` as well which e.g. allows to load a graph without having full checkouts of every package.

However, this is just a stepping stone towards that goal, since currently we are playing a bit fast-and-lose with file system objects during manifest loading (we're loading the same manifest from the git-backed fs, then directly via the local file system, for example).